requirements.txt
aws-cdk-lib
constructs>10.0.0
pillow
pyodata==1.5.0
requests==2.22.0
urllib3<2